+++ evalml/pipelines/components/transformers/imputers/target_imputer.py
@@ -47,6 +47,11 @@ class TargetImputer(Transformer, metaclass=TargetImputerMeta):
 
     name = "Target Imputer"
     hyperparameter_ranges = {"impute_strategy": ["mean", "median", "most_frequent"]}
+    """{
+        "impute_strategy": ["mean", "median", "most_frequent"]
+    }"""
+    modifies_features = False
+    modifies_target = True
 
     def __init__(
         self, impute_strategy="most_frequent", fill_value=None, random_seed=0, **kwargs