+++ evalml/pipelines/components/transformers/preprocessing/drop_null_columns.py
@@ -14,6 +14,7 @@ class DropNullColumns(Transformer):
 
     name = "Drop Null Columns Transformer"
     hyperparameter_ranges = {}
+    """{}"""
 
     def __init__(self, pct_null_threshold=1.0, random_seed=0, **kwargs):
         if pct_null_threshold < 0 or pct_null_threshold > 1: