+++ evalml/pipelines/components/transformers/preprocessing/polynomial_detrender.py
@@ -17,8 +17,10 @@ class PolynomialDetrender(TargetTransformer):
     """
 
     name = "Polynomial Detrender"
-
     hyperparameter_ranges = {"degree": Integer(1, 3)}
+    """{
+        "degree": Integer(1, 3)
+    }"""
 
     def __init__(self, degree=1, random_seed=0, **kwargs):
         if not isinstance(degree, int):