+++ evalml/pipelines/components/transformers/transformer.py
@@ -30,6 +30,9 @@ class Transformer(ComponentBase):
     """
 
     model_family = ModelFamily.NONE
+    """ModelFamily.NONE"""
+    modifies_features = True
+    modifies_target = False
 
     def transform(self, X, y=None):
         """Transforms data X.
@@ -85,6 +88,9 @@ class Transformer(ComponentBase):
 
 class TargetTransformer(Transformer):
     """A component that transforms the target."""
+
+    modifies_features = False
+    modifies_target = True
 
     @abstractmethod
     def inverse_transform(self, y):