+++ evalml/pipelines/pipeline_base.py
@@ -56,6 +56,7 @@ class PipelineBase(ABC, metaclass=PipelineBaseMeta):
     """
 
     problem_type = None
+    """None"""
 
     def __init__(
         self,