+++ evalml/tests/automl_tests/dask_tests/test_automl_dask.py
@@ -1,16 +1,20 @@
+from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
+
 import numpy as np
 import pytest
-from dask.distributed import Client, LocalCluster
+from dask import distributed as dd
+from dask.distributed import LocalCluster
 
 from evalml.automl import AutoMLSearch
 from evalml.automl.automl_algorithm import IterativeAlgorithm
 from evalml.automl.callbacks import raise_error_callback
-from evalml.automl.engine import DaskEngine, SequentialEngine
+from evalml.automl.engine import CFEngine, DaskEngine, SequentialEngine
+from evalml.automl.engine.cf_engine import CFClient
 from evalml.tests.automl_tests.dask_test_utils import (
-    TestPipelineFast,
-    TestPipelineSlow,
-    TestPipelineWithFitError,
-    TestPipelineWithScoreError,
+    DaskPipelineFast,
+    DaskPipelineSlow,
+    DaskPipelineWithFitError,
+    DaskPipelineWithScoreError,
 )
 from evalml.tuners import SKOptTuner
 
@@ -27,10 +31,65 @@ def cluster():
     dask_cluster.close()
 
 
-def test_automl(X_y_binary_cls, cluster, sequential_engine):
+@pytest.fixture(scope="module")
+def thread_pool():
+    pool = ThreadPoolExecutor()
+    yield pool
+    pool.shutdown()
+
+
+@pytest.fixture(scope="module")
+def process_pool():
+    pool = ProcessPoolExecutor()
+    yield pool
+    pool.shutdown()
+
+
+# List of tuples indicating which engine types and resource pools to parametrize tests over.
+# Removed '("CFEngine", "process")' as it doesn't work properly on GitHub test runners.
+engine_and_resource_types = [("CFEngine", "thread"), ("DaskEngine", "N/A")]
+
+
+def _get_engine_support(parallel_engine_type, thread_pool, cluster):
+    """Helper function to return the proper combination of resource pool, client class and
+    engine class for testing purposes.
+
+    e.g. The CFEngine can be run either with a ThreadPoolExecutor or a ProcessPoolExecutor,
+        so _get_engine_support("CFEngine", thread_pool, cluster) returns a
+        tuple of (ThreadPoolExecutor, cf.Client, cf.CFEngine)
+    """
+    if parallel_engine_type == "CFEngine":
+        resources = thread_pool
+        client_class = CFClient
+        engine_class = CFEngine
+    elif parallel_engine_type == "DaskEngine":
+        resources = cluster
+        client_class = dd.Client
+        engine_class = DaskEngine
+    return resources, client_class, engine_class
+
+
+@pytest.mark.parametrize(
+    "parallel_engine_type,pool_type",
+    engine_and_resource_types,
+)
+def test_automl(
+    parallel_engine_type,
+    pool_type,
+    X_y_binary_cls,
+    cluster,
+    thread_pool,
+    process_pool,
+    sequential_engine,
+):
     """Comparing the results of parallel and sequential AutoML to each other."""
-    with Client(cluster) as client:
-        parallel_engine = DaskEngine(client)
+
+    resources, client_class, engine_class = _get_engine_support(
+        parallel_engine_type, thread_pool, cluster
+    )
+
+    with client_class(resources) as client:
+        parallel_engine = engine_class(client)
         X, y = X_y_binary_cls
         par_automl = AutoMLSearch(
             X_train=X, y_train=y, problem_type="binary", engine=parallel_engine
@@ -62,12 +121,28 @@ def test_automl(X_y_binary_cls, cluster, sequential_engine):
         )
 
 
-def test_automl_max_iterations(X_y_binary_cls, cluster, sequential_engine):
+@pytest.mark.parametrize(
+    "parallel_engine_type,pool_type",
+    engine_and_resource_types,
+)
+def test_automl_max_iterations(
+    parallel_engine_type,
+    pool_type,
+    X_y_binary_cls,
+    cluster,
+    thread_pool,
+    process_pool,
+    sequential_engine,
+):
     """Making sure that the max_iterations parameter limits the number of pipelines run."""
 
     X, y = X_y_binary_cls
-    with Client(cluster) as client:
-        parallel_engine = DaskEngine(client)
+    resources, client_class, engine_class = _get_engine_support(
+        parallel_engine_type, thread_pool, cluster
+    )
+
+    with client_class(resources) as client:
+        parallel_engine = engine_class(client)
 
         max_iterations = 4
         par_automl = AutoMLSearch(
@@ -93,14 +168,31 @@ def test_automl_max_iterations(X_y_binary_cls, cluster, sequential_engine):
         assert len(sequential_rankings) == len(parallel_rankings) == max_iterations
 
 
-def test_automl_train_dask_error_callback(X_y_binary_cls, cluster, caplog):
+@pytest.mark.parametrize(
+    "parallel_engine_type,pool_type",
+    engine_and_resource_types,
+)
+def test_automl_train_dask_error_callback(
+    parallel_engine_type,
+    pool_type,
+    X_y_binary_cls,
+    cluster,
+    thread_pool,
+    process_pool,
+    sequential_engine,
+    caplog,
+):
     """Make sure the pipeline training error message makes its way back from the workers."""
     caplog.clear()
-    with Client(cluster) as client:
-        parallel_engine = DaskEngine(client)
-        X, y = X_y_binary_cls
+    X, y = X_y_binary_cls
+    resources, client_class, engine_class = _get_engine_support(
+        parallel_engine_type, thread_pool, cluster
+    )
 
-        pipelines = [TestPipelineWithFitError({})]
+    with client_class(resources) as client:
+        parallel_engine = engine_class(client)
+
+        pipelines = [DaskPipelineWithFitError({})]
         automl = AutoMLSearch(
             X_train=X,
             y_train=y,
@@ -114,14 +206,30 @@ def test_automl_train_dask_error_callback(X_y_binary_cls, cluster, caplog):
         assert "Train error for PipelineWithError: Yikes" in caplog.text
 
 
-def test_automl_score_dask_error_callback(X_y_binary_cls, cluster, caplog):
+@pytest.mark.parametrize(
+    "parallel_engine_type,pool_type",
+    engine_and_resource_types,
+)
+def test_automl_score_dask_error_callback(
+    parallel_engine_type,
+    pool_type,
+    X_y_binary_cls,
+    cluster,
+    thread_pool,
+    process_pool,
+    sequential_engine,
+    caplog,
+):
     """Make sure the pipeline scoring error message makes its way back from the workers."""
     caplog.clear()
-    with Client(cluster) as client:
-        parallel_engine = DaskEngine(client)
+    X, y = X_y_binary_cls
+    resources, client_class, engine_class = _get_engine_support(
+        parallel_engine_type, thread_pool, cluster
+    )
 
-        X, y = X_y_binary_cls
-        pipelines = [TestPipelineWithScoreError({})]
+    with client_class(resources) as client:
+        parallel_engine = engine_class(client)
+        pipelines = [DaskPipelineWithScoreError({})]
         automl = AutoMLSearch(
             X_train=X,
             y_train=y,
@@ -137,17 +245,34 @@ def test_automl_score_dask_error_callback(X_y_binary_cls, cluster, caplog):
         assert "Score error for PipelineWithError" in caplog.text
 
 
-def test_automl_immediate_quit(X_y_binary_cls, cluster, caplog):
+@pytest.mark.parametrize(
+    "parallel_engine_type,pool_type",
+    engine_and_resource_types,
+)
+def test_automl_immediate_quit(
+    parallel_engine_type,
+    pool_type,
+    X_y_binary_cls,
+    cluster,
+    thread_pool,
+    process_pool,
+    sequential_engine,
+    caplog,
+):
     """Make sure the AutoMLSearch quits when error_callback is defined and does no further work."""
     caplog.clear()
     X, y = X_y_binary_cls
-    with Client(cluster) as client:
-        parallel_engine = DaskEngine(client)
+    resources, client_class, engine_class = _get_engine_support(
+        parallel_engine_type, thread_pool, cluster
+    )
+
+    with client_class(resources) as client:
+        parallel_engine = engine_class(client)
 
         pipelines = [
-            TestPipelineFast({}),
-            TestPipelineWithFitError({}),
-            TestPipelineSlow({}),
+            DaskPipelineFast({}),
+            DaskPipelineWithFitError({}),
+            DaskPipelineSlow({}),
         ]
         automl = AutoMLSearch(
             X_train=X,
@@ -178,9 +303,9 @@ def test_automl_immediate_quit(X_y_binary_cls, cluster, caplog):
 
         # Make sure the automl algorithm stopped after the broken pipeline raised
         assert len(automl.full_rankings) < len(pipelines)
-        assert TestPipelineSlow.custom_name not in set(
+        assert DaskPipelineSlow.custom_name not in set(
             automl.full_rankings["pipeline_name"]
         )
-        assert TestPipelineWithFitError.custom_name not in set(
+        assert DaskPipelineWithFitError.custom_name not in set(
             automl.full_rankings["pipeline_name"]
         )