+++ evalml/tests/model_understanding_tests/test_permutation_importance.py
@@ -660,3 +660,14 @@ def test_get_permutation_importance_one_column_pipeline_does_not_support_fast(
         calculate_permutation_importance_one_column(
             pipeline_class(params), X, y, 0, "log loss binary", fast=True
         )
+
+
+def test_permutation_importance_unknown(X_y_binary):
+    # test to see if we can get permutation importance fine with a dataset that has unknown features
+    X, y = X_y_binary
+    X = pd.DataFrame(X)
+    X.ww.init(logical_types={0: "unknown"})
+    pl = BinaryClassificationPipeline(["Random Forest Classifier"])
+    pl.fit(X, y)
+    s = calculate_permutation_importance(pl, X, y, objective="Log Loss Binary")
+    assert not s.isnull().any().any()